Fix some issues and put tests in proper places
