+++ crates/flux-fhir-analysis/src/wf/param_usage.rs
@@ -120,7 +120,7 @@ impl<'a, 'genv, 'tcx> ParamUsesChecker<'a, 'genv, 'tcx> {
             }
             fhir::ExprKind::Constructor(_path, exprs, _spread) => {
                 for expr in exprs {
-                    self.check_func_params_uses(&expr.expr, is_top_level_conj, is_top_level_var);
+                    self.check_func_params_uses(&expr.expr, false, false);
                 }
             }
         }
