+++ tests/tests/neg/constructors/test01.rs
@@ -1,3 +1,5 @@
+#![allow(unused)]
+
 #[flux::refined_by(x: int, y: int)]
 pub struct X {
     #[flux::field(u32[x])]
@@ -6,9 +8,22 @@ pub struct X {
     y: u32,
 }
 
-#[flux::sig(fn (x: X[@old_x]) -> X[X { y: true, x: 1 }])] //~ ERROR mismatched sorts
-fn f1(mut x: X) -> X {
+#[flux::refined_by(x: X)]
+pub struct Y {
+    #[flux::field(X[x])]
+    big_x: X,
+}
+
+#[flux::sig(fn (x: X[@old_x]) -> X[X { y: 2, x: 1 }])]
+fn f(mut x: X) -> X {
     x.x = 1;
     x.y = 2;
     x
+}
+
+#[flux::sig(fn (x: Y[@old_y]) -> Y[Y { x: X { x: 1, y: 1 } }])]
+fn f_through_y(mut y: Y) -> Y {
+    y.big_x.x = 1;
+    y.big_x.y = 2;
+    y //~ ERROR refinement type error
 }
